Move and rename xtra_libp2p_ping::ping submodule
