+++ xtra-libp2p-ping/src/lib.rs
@@ -15,6 +15,8 @@ use xtra_libp2p::OpenSubstream;
 use xtra_productivity::xtra_productivity;
 use xtras::SendInterval;
 
+mod protocol;
+
 /// The name of the official ipfs/libp2p ping protocol.
 ///
 /// Using this indicates that we are wire-compatible with other libp2p/ipfs nodes.
@@ -106,7 +108,7 @@ impl Actor {
                     let stream = endpoint
                         .send(OpenSubstream::single_protocol(peer, PROTOCOL_NAME))
                         .await??;
-                    let latency = ping::send(stream).await?;
+                    let latency = protocol::send(stream).await?;
 
                     this.send(RecordLatency {
                         peer,
@@ -143,7 +145,7 @@ impl Actor {
     async fn handle(&mut self, message: NewInboundSubstream) {
         let NewInboundSubstream { stream, peer } = message;
 
-        let future = ping::recv(stream);
+        let future = protocol::recv(stream);
 
         self.tasks.add_fallible(future, move |e| async move {
             tracing::debug!(%peer, "Inbound ping protocol failed: {e}");
@@ -170,54 +172,6 @@ static PEER_LATENCY_HISTOGRAM: Lazy<Histogram> = Lazy::new(|| {
     )
     .unwrap()
 });
-
-/// The actual protocol functions for sending ping messages.
-///
-/// Insbired by https://github.com/libp2p/rust-libp2p/blob/102509afe3a3b984e43a88dbe4de935fde36f319/protocols/ping/src/protocol.rs#L82-L113.
-mod ping {
-    pub const SIZE: usize = 32;
-
-    use futures::AsyncReadExt;
-    use futures::AsyncWriteExt;
-    use rand::distributions;
-    use rand::thread_rng;
-    use rand::Rng;
-    use std::io;
-    use std::time::Duration;
-    use std::time::Instant;
-    use xtra_libp2p::Substream;
-
-    /// Sends a ping and waits for the pong.
-    pub async fn send(mut stream: Substream) -> io::Result<Duration> {
-        let payload: [u8; SIZE] = thread_rng().sample(distributions::Standard);
-        stream.write_all(&payload).await?;
-        stream.flush().await?;
-
-        let started = Instant::now();
-
-        let mut recv_payload = [0u8; SIZE];
-        stream.read_exact(&mut recv_payload).await?;
-
-        if recv_payload != payload {
-            return Err(io::Error::new(
-                io::ErrorKind::InvalidData,
-                "Ping payload mismatch",
-            ));
-        }
-
-        Ok(started.elapsed())
-    }
-
-    /// Waits for a ping and sends a pong.
-    pub async fn recv(mut stream: Substream) -> io::Result<()> {
-        let mut payload = [0u8; SIZE];
-        stream.read_exact(&mut payload).await?;
-        stream.write_all(&payload).await?;
-        stream.flush().await?;
-
-        Ok(())
-    }
-}
 
 #[cfg(test)]
 mod tests {
